+++ CHANGELOG.md
@@ -1143,6 +1143,7 @@ activity and on its child transactions (which can be actuals, refunds, and adjus
 - Enable activity linking for ISPF activities
 - ISPF activities can have multiple ISPF themes
 - ISPF activities can be tagged via the UI - tags can only be chosen from a BEIS-defined codelist
+- Allow users to designate an ISPF activity as having no ISPF partner countries
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-123...HEAD
 [release-123]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-122...release-123